Allows users to provide their own S3 bucket for use in holding LB Access Logs
Our organization has a single bucket designated for LB access logs in
each environment, and we'd like to avoid having a 1:1 ratio between LBs
and log buckets.

This commit enables the module user to provide the ID of a pre-existing
S3 bucket that the LB is configured to use.  This becomes mutually
exclusive with the existing "enable_s3_logs" boolean.